refactor: infra基本モジュールのテスト分離(compute,logging,parse)
- tests/infra/test_{compute,logging,parse}.pyを作成
- src/infra/{compute,logging,parse}.pyからTestModクラスを削除
- tests/conftest.pyにmockディレクトリマッピング追加
- pyfakefs対応でtest_parse.pyを修正
- 全テスト通過を確認（70 passed）
