+++ src/infra/parse.py
@@ -72,87 +72,3 @@ def _extract_local_name(tag: str) -> str:
     return tag
 
 
-class TestMod:
-    def test_parse_rss_returns_rss_root(self) -> None:
-        """
-        docs:
-            目的:
-                RSSモックファイルを解析し、rssルート要素が取得できることを確認する。
-            検証観点:
-                - parse_rss の戻り値が Element インスタンスであること。
-                - ルートタグが rss であること。
-                - channel 要素が存在すること。
-        """
-
-        fixture_path = Path(__file__).resolve().parents[2] / "mock" / "google_news.rss"
-        content = fixture_path.read_bytes()
-
-        root = parse_rss(content)
-
-        assert isinstance(root, Element)
-        assert _extract_local_name(root.tag) == "rss"
-        channel = root.find("channel")
-        assert channel is not None
-
-    def test_parse_rss_rejects_non_rss_root(self) -> None:
-        """
-        docs:
-            目的:
-                rss以外のルート要素を持つXMLを渡した際に例外を送出することを確認する。
-            検証観点:
-                - ValueError が送出される。
-        """
-
-        invalid_xml = "<feed><title>Example</title></feed>"
-
-        with pytest.raises(ValueError):
-            parse_rss(invalid_xml)
-
-    def test_parse_rss_rejects_empty_payload(self) -> None:
-        """
-        docs:
-            目的:
-                空文字列を渡した場合に適切な例外が送出されることを確認する。
-            検証観点:
-                - ValueError が送出される。
-        """
-
-        with pytest.raises(ValueError):
-            parse_rss("")
-
-    def test_parse_rss_rejects_broken_xml(self) -> None:
-        """
-        docs:
-            目的:
-                閉じタグの欠落したXMLを渡した場合に解析エラーとなることを確認する。
-            検証観点:
-                - ValueError が送出される。
-        """
-
-        broken_xml = "<rss><channel><title>Example"  # 閉じタグ欠落
-
-        with pytest.raises(ValueError):
-            parse_rss(broken_xml)
-
-    def test_parse_rss_sanitizes_unescaped_ampersand(self) -> None:
-        """
-        docs:
-            目的:
-                アンパサンドがエスケープされていない入力でも解析できることを確認する。
-            検証観点:
-                - `&` が含まれていても ValueError にならない。
-                - channel 要素を取得できる。
-        """
-
-        xml_with_ampersand = """
-            <rss version="2.0">
-                <channel>
-                    <title>Example & Co.</title>
-                </channel>
-            </rss>
-        """
-
-        root = parse_rss(xml_with_ampersand)
-
-        assert _extract_local_name(root.tag) == "rss"
-        assert root.find("channel") is not None
